+++ src/ts/ui/index.ts
@@ -90,9 +90,14 @@ export class Ui {
 
         // set default value
         let initValue = localStorage.getItem("vditor" + vditor.id);
-        if ((!vditor.options.cache || !initValue) && vditor.originalInnerHTML) {
-            initValue = await html2md(vditor, vditor.originalInnerHTML);
+        if (!vditor.options.cache || !initValue) {
+            if (vditor.options.value) {
+                initValue = vditor.options.value;
+            } else if (vditor.originalInnerHTML) {
+                initValue = await html2md(vditor, vditor.originalInnerHTML);
+            }
         }
+
         if (!initValue) {
             return;
         }